requirements.txt
django
pylint
autopep8
mysqlclient==2.1.1
mysql-connector-python